+++ bftengine/src/bftengine/ReplicaImp.hpp
@@ -43,8 +43,8 @@
 #include "FakeClock.hpp"
 #include <ccron/ticks_generator.hpp>
 #include "EpochManager.hpp"
-#include "BftExecutionEngineBase.hpp"
 #include "PerfMetrics.hpp"
+#include "BftExecutionEngine.hpp"
 
 namespace preprocessor {
 class PreProcessResultMsg;
@@ -260,8 +260,10 @@ class ReplicaImp : public InternalReplicaApi, public ReplicaForStateTransfer {
   GaugeHandle accumulating_batch_avg_time_;
   GaugeHandle deferredRORequestsMetric_;
   GaugeHandle deferredMessagesMetric_;
+
   // The first commit path being attempted for a new request.
   StatusHandle metric_first_commit_path_;
+  CounterHandle metric_number_of_current_executions_;
   CounterHandle batch_closed_on_logic_off_;
   CounterHandle batch_closed_on_logic_on_;
   CounterHandle metric_indicator_of_non_determinism_;
@@ -584,6 +586,7 @@ class ReplicaImp : public InternalReplicaApi, public ReplicaForStateTransfer {
   void addTimers();
   void startConsensusProcess(PrePrepareMsg* pp, bool isCreatedEarlier);
   void startConsensusProcess(PrePrepareMsg* pp);
+  void postBftExecutionActions(PrePrepareMsg* ppMsg, IRequestsHandler::ExecutionRequestsQueue& requests_for_execution);
   /**
    * Updates both seqNumInfo and slow_path metric
    * @param seqNumInfo
